requirements.txt
fastapi
uvicorn
docxtpl
python-docx
qrcode[pil]
Pillow
requests
boto3
python-dotenv

